requirements.txt
pyzmq
msgpack-python
jsonschema
termcolor
semver
click
progressbar2
aiohttp
aiodns
cchardet
aiohttp-cors
pygit2
